rename file to mach hook name
