+++ src/hooks/use-computation-results-count.ts
